Added an "Are you sure?" prompt to the VM-updater script.

+++ updateVirtualAGC-VM64.sh
@@ -3,10 +3,14 @@
 # by getting the latest version of the source tree and building it.  However,
 # it can potentially perform other tasks as well.
 
-cd ~/git/virtualagc
-git stash drop
-git pull
-time make clean install clean
-echo "Terminated.  Hit ENTER key or close this window."
-read
+read -p "Are you sure (y/N)? " -n 1 -r
+if [[ "$REPLY" =~ ^[Yy]$ ]]
+then
+        cd ~/git/virtualagc
+        git stash drop
+        git pull
+        time make clean install clean
+        echo "Terminated.  Hit ENTER key or close this window."
+        read
+fi
